Addressed various Clang tidy concerns

+++ firmware/include/modes/CircleMode.h
@@ -12,10 +12,10 @@ private:
     static constexpr float y = (GRID_ROWS - 1) / 2.0F;
 
     static inline const uint8_t maxRadius{
-        3 + ceilf((max(GRID_COLUMNS * PITCH_HORIZONTAL / static_cast<float>(PITCH_VERTICAL),
-                       GRID_ROWS *PITCH_VERTICAL / static_cast<float>(PITCH_HORIZONTAL)) /
-                   std::numbers::sqrt2_v<float> * 3.0F / 2.0F) /
-                  2.0F)};
+        3 + static_cast<uint8_t>(ceilf((max(GRID_COLUMNS * PITCH_HORIZONTAL / static_cast<float>(PITCH_VERTICAL),
+                                            GRID_ROWS *PITCH_VERTICAL / static_cast<float>(PITCH_HORIZONTAL)) /
+                                        std::numbers::sqrt2_v<float> * 3.0F / 2.0F) /
+                                       2.0F))};
 
     bool lit = true;
